requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
pyvisa==1.15.0
requests==2.32.3
vncdotool==1.2.0
Pillow>=9.0.0
